[REFACT] 예약 서비스 인터페이스 삭제
